DDF-2507 Added class to handle attribute registration for XmlInputTransformer
