+++ catalog/transformer/catalog-transformer-streaming-lib/src/test/java/org/codice/ddf/transformer/xml/streaming/lib/TestGenericXmlLib.java
@@ -19,6 +19,7 @@ import static org.hamcrest.Matchers.notNullValue;
 import static org.junit.Assert.assertThat;
 import static org.junit.Assert.fail;
 import static org.mockito.Matchers.any;
+import static org.mockito.Mockito.doReturn;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
@@ -33,8 +34,8 @@ import java.util.Set;
 
 import org.codice.ddf.transformer.xml.streaming.SaxEventHandler;
 import org.codice.ddf.transformer.xml.streaming.SaxEventHandlerFactory;
+import org.junit.Before;
 import org.junit.Test;
-import org.osgi.framework.BundleContext;
 import org.xml.sax.SAXException;
 
 import ddf.catalog.data.Attribute;
@@ -49,6 +50,14 @@ import ddf.catalog.transform.CatalogTransformerException;
 
 public class TestGenericXmlLib {
 
+    MetacardTypeRegister mockMetacardTypeRegister;
+
+    @Before
+    public void setup() {
+        mockMetacardTypeRegister = mock(MetacardTypeRegister.class);
+        doReturn(BasicTypes.BASIC_METACARD).when(mockMetacardTypeRegister).getMetacardType();
+    }
+
     @Test
     public void testNormalTransform() throws FileNotFoundException, CatalogTransformerException {
 
@@ -57,16 +66,13 @@ public class TestGenericXmlLib {
         when(saxEventHandlerFactory.getId()).thenReturn("test");
         SaxEventHandler handler = getNewHandler();
         when(saxEventHandlerFactory.getNewSaxEventHandler()).thenReturn(handler);
-        XmlInputTransformer xmlInputTransformer = new XmlInputTransformer() {
-            @Override
-            public BundleContext getContext() {
-                return mock(BundleContext.class);
-            }
-        };
+        XmlInputTransformer xmlInputTransformer = new XmlInputTransformer();
         xmlInputTransformer.setSaxEventHandlerConfiguration(Collections.singletonList("test"));
         xmlInputTransformer.setSaxEventHandlerFactories(Collections.singletonList(
                 saxEventHandlerFactory));
-        assertThat(xmlInputTransformer.getMetacardType(), is(notNullValue()));
+        xmlInputTransformer.setDynamicMetacardTypeRegister(mockMetacardTypeRegister);
+        assertThat(xmlInputTransformer.getDynamicMetacardTypeRegister()
+                .getMetacardType(), is(notNullValue()));
         Metacard metacard = null;
         try {
             metacard = xmlInputTransformer.transform(inputStream, "test");
@@ -88,15 +94,11 @@ public class TestGenericXmlLib {
         when(saxEventHandlerFactory.getId()).thenReturn("test");
         SaxEventHandler handler = getNewHandler();
         when(saxEventHandlerFactory.getNewSaxEventHandler()).thenReturn(handler);
-        XmlInputTransformer xmlInputTransformer = new XmlInputTransformer() {
-            @Override
-            public BundleContext getContext() {
-                return mock(BundleContext.class);
-            }
-        };
+        XmlInputTransformer xmlInputTransformer = new XmlInputTransformer();
         xmlInputTransformer.setSaxEventHandlerConfiguration(Collections.singletonList("test"));
         xmlInputTransformer.setSaxEventHandlerFactories(Collections.singletonList(
                 saxEventHandlerFactory));
+        xmlInputTransformer.setDynamicMetacardTypeRegister(mockMetacardTypeRegister);
         try {
             xmlInputTransformer.transform(inputStream, "test");
         } catch (IOException e) {
@@ -111,12 +113,7 @@ public class TestGenericXmlLib {
         when(saxEventHandlerFactory.getId()).thenReturn("test");
         SaxEventHandler handler = getNewHandler();
         when(saxEventHandlerFactory.getNewSaxEventHandler()).thenReturn(handler);
-        XmlInputTransformer xmlInputTransformer = new XmlInputTransformer() {
-            @Override
-            public BundleContext getContext() {
-                return mock(BundleContext.class);
-            }
-        };
+        XmlInputTransformer xmlInputTransformer = new XmlInputTransformer();
         xmlInputTransformer.setSaxEventHandlerConfiguration(Collections.singletonList("test"));
         xmlInputTransformer.setSaxEventHandlerFactories(Collections.singletonList(
                 saxEventHandlerFactory));
@@ -134,30 +131,24 @@ public class TestGenericXmlLib {
         when(saxEventHandlerFactory.getId()).thenReturn("test");
         SaxEventHandler handler = getNewHandler();
         when(saxEventHandlerFactory.getNewSaxEventHandler()).thenReturn(handler);
-        XmlInputTransformer xmlInputTransformer = new XmlInputTransformer() {
-            @Override
-            public BundleContext getContext() {
-                return mock(BundleContext.class);
-            }
-        };
+        XmlInputTransformer xmlInputTransformer = new XmlInputTransformer();
         xmlInputTransformer.setSaxEventHandlerFactories(Collections.singletonList(
                 saxEventHandlerFactory));
-        MetacardType metacardType = xmlInputTransformer.getMetacardType();
+        xmlInputTransformer.setDynamicMetacardTypeRegister(mockMetacardTypeRegister);
+        MetacardType metacardType = xmlInputTransformer.getDynamicMetacardTypeRegister()
+                .getMetacardType();
         assertThat(metacardType.getAttributeDescriptors(),
                 is(BasicTypes.BASIC_METACARD.getAttributeDescriptors()));
     }
 
     @Test
     public void testNoFactoriesTransform() {
-        XmlInputTransformer xmlInputTransformer = new XmlInputTransformer() {
-            @Override
-            public BundleContext getContext() {
-                return mock(BundleContext.class);
-            }
-        };
+        XmlInputTransformer xmlInputTransformer = new XmlInputTransformer();
         xmlInputTransformer.setSaxEventHandlerConfiguration(Collections.singletonList("test"));
         xmlInputTransformer.setSaxEventHandlerFactories(null);
-        MetacardType metacardType = xmlInputTransformer.getMetacardType();
+        xmlInputTransformer.setDynamicMetacardTypeRegister(mockMetacardTypeRegister);
+        MetacardType metacardType = xmlInputTransformer.getDynamicMetacardTypeRegister()
+                .getMetacardType();
         assertThat(metacardType.getAttributeDescriptors(),
                 is(BasicTypes.BASIC_METACARD.getAttributeDescriptors()));
     }
@@ -192,16 +183,13 @@ public class TestGenericXmlLib {
         SaxEventHandler handler = getNewHandler();
 
         when(saxEventHandlerFactory.getNewSaxEventHandler()).thenReturn(handler);
-        XmlInputTransformer xmlInputTransformer = new XmlInputTransformer() {
-            @Override
-            public BundleContext getContext() {
-                return mock(BundleContext.class);
-            }
-        };
+        XmlInputTransformer xmlInputTransformer = new XmlInputTransformer();
         xmlInputTransformer.setSaxEventHandlerConfiguration(Collections.singletonList("test"));
         xmlInputTransformer.setSaxEventHandlerFactories(Collections.singletonList(
                 saxEventHandlerFactory));
-        assertThat(xmlInputTransformer.getMetacardType(), is(notNullValue()));
+        xmlInputTransformer.setDynamicMetacardTypeRegister(mockMetacardTypeRegister);
+        assertThat(xmlInputTransformer.getDynamicMetacardTypeRegister()
+                .getMetacardType(), is(notNullValue()));
         Metacard metacard = null;
         try {
             metacard = xmlInputTransformer.transform(inputStream, "test");
